issue #129: move packages to internal/app
