+++ internal/app/get/get.go
@@ -1,8 +1,8 @@
 package details
 
 import (
-	"github.com/jbrekelmans/kube-compose/internal/pkg/k8smeta"
-	"github.com/jbrekelmans/kube-compose/pkg/config"
+	"github.com/jbrekelmans/kube-compose/internal/app/config"
+	"github.com/jbrekelmans/kube-compose/internal/app/k8smeta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	clientV1 "k8s.io/client-go/kubernetes/typed/core/v1"
