+++ internal/app/k8smeta/k8smeta_test.go
@@ -3,7 +3,7 @@ package k8smeta
 import (
 	"testing"
 
-	"github.com/jbrekelmans/kube-compose/pkg/config"
+	"github.com/jbrekelmans/kube-compose/internal/app/config"
 	dockerComposeConfig "github.com/jbrekelmans/kube-compose/pkg/docker/compose/config"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
