+++ internal/app/up/status.go
