+++ internal/app/up/status_test.go
